requirements.txt
chromadb
google.generativeai
langchain
langchain-google-genai
langchain_community
jupyter
pypdf
python-dotenv
yandex-chain
nbformat 
nbconvert
telegram
python-telegram-bot
unstructured
markdown
unstructured_inference
pillow_heif
sentence-transformers